feat(path/encoding): attempt to decode paths
This commit attempts to handle requests for file paths that include
percent encodings.

A real world example of this is a Next.js client side router, which is
trying to make requests such as:

GET /out/_next/static/chunks/pages/editor/%2F%5BtemplateId%5D-2cead6d6c9c5e605.js

This request should be resolved to

 out/_next/static/chunks/pages/editor/[templateId]-2cead6d6c9c5e605.js

However, the current implementation does not -- and either returns a
404, or if a fallback was configured, returns it.

This commit should not be merged as-is, since it disregards the existing
PATH_INFO header.
I also don't have enough context to understand how to reconcile those.

So opening this as a draft in the hope that somoeone with more context
can help.

This is currently a blocker for using Next.js based SPAs.

Signed-off-by: Radu Matei <radu.matei@fermyon.com>